Allow use as trait, rather than via explicit inheritance.
